refactor: split report/mod.rs into query, display, export submodules
